refactored nodes into smaller modules
