+++ src/nodes/common/__init__.py
@@ -1,0 +1,23 @@
+"""
+Common utilities for all nodes.
+
+This package contains shared functionality that is used across different nodes
+to avoid code duplication and maintain consistency.
+"""
+
+from .llm_utils import load_model, create_messages, extract_response_content
+from .state_utils import build_error_state, apply_updates_to_investigations
+from .response_processing import (
+    process_structured_response,
+    ensure_proper_format,
+)
+
+__all__ = [
+    "load_model",
+    "create_messages",
+    "extract_response_content",
+    "build_error_state",
+    "apply_updates_to_investigations",
+    "process_structured_response",
+    "ensure_proper_format",
+]
