+++ src/nodes/planner/__init__.py
@@ -1,0 +1,31 @@
+"""
+Planner Node.
+
+This module orchestrates the planning workflow by loading plans,
+selecting appropriate plans, and updating investigations with planning results.
+"""
+
+# Import the main node function from core
+from .core import planner_node
+
+# Import supporting functions for backwards compatibility with tests
+from .planning import (
+    load_available_plans,
+    execute_plan_selection,
+    process_planning_response,
+    PlanningResponse,
+    DevicePlan,
+)
+from util.plans import plans_to_string, load_plans
+from .context import extract_investigations_summary
+from .state import build_successful_planning_state, build_failed_planning_state
+from nodes.common import load_model
+
+# Aliases for backwards compatibility with tests
+_load_available_plans = load_available_plans
+_load_model = lambda: load_model()
+_execute_plan_selection = execute_plan_selection
+_process_planning_response = process_planning_response
+_extract_investigations_summary = extract_investigations_summary
+_build_successful_planning_state = build_successful_planning_state
+_build_failed_planning_state = build_failed_planning_state
